# expression: implement vectorized evaluation for `builtinSubStringAndStringSig`